Move support code into a better place
